Add nominal conditions to Network class
This introduces nominal conditions (m_flow ans c_p) to the Network
class. These are used at initialization to initialize the values of
the Network coefficients.

This is needed to simplify the interface to the new gFunction class, so
that m_flow and c_p are not needed to evaluate the g-function.