Add Solver classes
The new Solver classes will be called by the gFunction class to
calculate the gFunction. This will facilitate the implementation of new
calculation methods as they are developed in the literature.

The _BaseSolver class implements the `solve` function, which is aimed
to be applicable independent of the solution method. Method-specific
tasks can be implemented in their own classes.