Add visualization features
The gFunction class can now plot the borehole temperatures and heat
extraction rates (both as a function of time and the profiles over the
length). These are often shown in publications and are useful to
interpret the result of g-function calculations.

In the past, modified versions of pygfunction were used to output these
for publications. The gFunction class and the new visualization methods
add this functionality natively while limiting the memory use through
the `profiles` options.