Add pure water functionality to Fluid object
Coolprop provides access to incompresible mass-based binary
mixtures[1]. The string for the fluid mixture [2] appears as follows:
```
'INCOMP::Mixture-Percent%'
```
Where Mixture is replaced with a mixture to mix in with water [1] and
Percent is replaced with a value from 0-100. Although, there is an xmin
and xmax value for each mixture (shown in mixture table at [1]).

For pure water properties, the string becomes `'Water'` [3].

closes MassimoCimmino#75

[1]: http://www.coolprop.dreamhosters.com:8010/binaries/sphinx/fluid_properties/Incompressibles.html#pure
[2]: http://www.coolprop.dreamhosters.com:8010/binaries/sphinx/fluid_properties/Incompressibles.html#mixture-examples
[3]: http://www.coolprop.dreamhosters.com:8010/binaries/sphinx/fluid_properties/Incompressibles.html#mixture-examples