Add second borehole thermal resistance function
It is not yet clear if this second borehole thermal resistance function
will be necessary. I will not know until I have completed this and
Massimo reviews it to see if he could use his own techniques to make it
happen.

I plan to make use of this function to calculate the effective borehole
thermal resistance for a concentric (coaxial) pipe arrangement.