Initial implementation
`Network._network_coefficients_from_pipe_coefficients` is not very
efficient due to a triple `for` loop to evaluate matrix coefficients.
On top of this, it is very hard to read and understand, even alongside
the equations from the reference paper (Cimmino, 2019).

An alternate method is used here, starting from the coefficients of the
equation [Desired_variable(borehole)] = [a]*T_{in,borehole} + [b]*[T_b]
and taking into account the borehole connectivity to reduce to
[Desired_variable(borehole)] = [c]*T_{in,network} + [d]*[T_b].

The result is more efficient and much easier to read.