Reformat variable names for m_flow
This updates the variable names for `m_flow` in the `pipes`, `networks`,
and `gfunction` modules. The variable name is now substituted with one of
`m_flow_pipe`, `m_flow_borehole` or `m_flow_network` depending on which
one applies. This attempts to clarify which value is expected as an input
to the functions and class methods. In many cases, there was some
confusion about the value to be provided.

Examples that make use of these modules are also updated.