Modify functions to accept unequal segment lengths
The segments function in boreholes.py had previously only accounted for
the segment lengths to be equal on a specific borehole. This
modification provides the ability for the segments to have different
lengths.

The gfunction.py file needed to address unequal segment lengths in the
_map_axial_segment_pairs function so that the boreholes.segments
function could properly create bore segments. The _baseSolver object
had updates to the segment_lengths(), borehole_segments() and
_check_inputs() functions.

  - The summation of the segment lengths for each borehole are
    asserted to be within a tenth of a meter of the borehole height
  - The borehole_segments() function accesses the new variable
    segmentLengths (if its not None) to split boreholes into
    segments
  - The segment_lengths() function flattens the 2D segmentLengths list
    if segmentLengths is not None