Refactor custom_borehole.py
The definition of each pipe type now has its own section in the script.
The mass flow is prescribed instead of the volume flow rate as is done in
the other examples. A mistake was corrected in the call to
`borehole_thermal_resistance()`: the borehole mass flow rate should be
provided (as opposed to the pipe mass flow rate). Pipes in single and
double U-tube boreholes are moved closer to the borehole wall and the
fluid is changed to propylene glycol 20%.